Add integration substeps to improve performance
Memory allocation and usage was placing a practical lower
limit to the integration step size. Integration substeps were
implemented that use the same integration method as before,
but will not remember the intermediate results. During testing,
the allowed for roughly a 100 fold reduction in time step
duration. Sub time steps of 10µs are now used by default.

This improvement in integration error has revealed a clear
chain fountain. It is quite fun to vary the parameters to
see the results.

It would be nice to have the simulation go for longer.
Perhaps a respawn mechanism will help.