Recycle beads by respawning beads on the floor
Also fix parameters open/close UI feature.

The bead recycling allows for a lower number of total beads which
makes the integration progress much quicker. Beads are only removed
from the floor if more than 20 are on the floor. This is to make sure
that the removal of the bead has insignificant effect on the chain
fountain. New beads are spawned at the same y-coordinate as the last
bead in the chain with the x coordinate offset by the link length
in the direction that keeps the spawning point within the x bounds
of the beaker. If there is not sufficient total beads, it is possible
for the spawning point to lift off the beaker or drop below it.
This is ok as it shows what happens when the beaker is not having an
effect on the chain physics. The result is that the chain fountain
dissipates.